.mul_agents/requirements.txt
camel-ai[all]==0.2.15a0
pygame>=2.5.2

fastapi
uvicorn
pyaudio
